move config file loading into a module
